impl, transaction for delete, and try...catch

+++ src/Laravel9TaskList/app/Http/Controllers/FolderController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Support\Facades\Auth;
-use App\Http\Requests\EditFolder;
-use App\Http\Requests\CreateFolder;
-use App\Models\Folder;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
+use App\Models\Folder;
+use App\Http\Requests\CreateFolder;
+use App\Http\Requests\EditFolder;
 
 class FolderController extends Controller
 {
@@ -18,15 +20,15 @@ class FolderController extends Controller
      */
     public function showCreateForm()
     {
-        // ログインユーザーに紐づくフォルダだけを取得
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $user->folders;
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $user->folders;
 
-        return view('folders/create');
-        // ?
-        // chpt.12, ユーザーに紐づけてフォルダを作成する
-        // return view('folders/create', compact('folders'));
+            return view('folders/create');
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in showCreateForm: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -39,17 +41,20 @@ class FolderController extends Controller
      */
     public function create(CreateFolder $request)
     {
+        try {
+            $folder = new Folder();
+            $folder->title = $request->title;
 
-        $folder = new Folder();
-        $folder->title = $request->title;
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $user->folders()->save($folder);
 
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $user->folders()->save($folder);
-
-        return redirect()->route('tasks.index', [
-            'folder' => $folder->id,
-        ]);
+            return redirect()->route('tasks.index', [
+                'folder' => $folder->id,
+            ]);
+        } catch (\Exception $e) {
+            Log::error('Error FolderController in create: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -61,14 +66,18 @@ class FolderController extends Controller
      */
     public function showEditForm(Folder $folder)
     {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $folder = $user->folders()->findOrFail($folder->id);
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $folder = $user->folders()->findOrFail($folder->id);
 
-        return view('folders/edit', [
-            'folder_id' => $folder->id,
-            'folder_title' => $folder->title,
-        ]);
+            return view('folders/edit', [
+                'folder_id' => $folder->id,
+                'folder_title' => $folder->title,
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in showEditForm: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -81,15 +90,19 @@ class FolderController extends Controller
      */
     public function edit(Folder $folder, EditFolder $request)
     {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $folder = $user->folders()->findOrFail($folder->id);
-        $folder->title = $request->title;
-        $folder->save();
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $folder = $user->folders()->findOrFail($folder->id);
+            $folder->title = $request->title;
+            $folder->save();
 
-        return redirect()->route('tasks.index', [
-            'folder' => $folder->id,
-        ]);
+            return redirect()->route('tasks.index', [
+                'folder' => $folder->id,
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in edit: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -101,14 +114,18 @@ class FolderController extends Controller
      */
     public function showDeleteForm(Folder $folder)
     {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $folder = $user->folders()->findOrFail($folder->id);
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $folder = $user->folders()->findOrFail($folder->id);
 
-        return view('folders/delete', [
-            'folder_id' => $folder->id,
-            'folder_title' => $folder->title,
-        ]);
+            return view('folders/delete', [
+                'folder_id' => $folder->id,
+                'folder_title' => $folder->title,
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error in showDeleteForm: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -120,17 +137,25 @@ class FolderController extends Controller
      */
     public function delete(Folder $folder)
     {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $folder = $user->folders()->findOrFail($folder->id);
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $folder = $user->folders()->findOrFail($folder->id);
 
-        $folder->tasks()->delete();
-        $folder->delete();
+            $folder = DB::transaction(function () use ($folder) {
+                if ($folder) throw new \Exception('500');
+                $folder->tasks()->delete();
+                $folder->delete();
+                return $folder;
+            });
 
-        $folder = Folder::first();
+            $folder = Folder::first();
 
-        return redirect()->route('tasks.index', [
-            'folder' => $folder->id
-        ]);
+            return redirect()->route('tasks.index', [
+                'folder' => $folder->id
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in delete: ' . $e->getMessage());
+        }
     }
 }
